+++ tests/xharness/Jenkins/Reports/IReportWriter.cs
@@ -2,7 +2,7 @@
 using System.IO;
 using Microsoft.DotNet.XHarness.iOS.Shared.Tasks;
 
-namespace Xharness.Jenkins {
+namespace Xharness.Jenkins.Reports {
 
 	/// <summary>
 	/// To be implemented by those classes that write reports regarding the results of the executed tasks.
